% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ci_tree_terminal_mean.R
\name{ci_collect_notes}
\alias{ci_collect_notes}
\title{Collect tuning notes}
\usage{
ci_collect_notes(x, ...)
}
\arguments{
\item{x}{A tuning object returned by \code{\link[=tune_ci_tree]{tune_ci_tree()}} or \code{\link[=tune_ci_forest]{tune_ci_forest()}}.}

\item{...}{Reserved for future extensions.}
}
\value{
A \code{data.table} with warnings and errors captured during tuning.
}
\description{
Collect tuning notes
}
