% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_parsnip_formula.R
\name{ci_gain}
\alias{ci_gain}
\title{Concentration-index gain metric}
\usage{
ci_gain(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A yardstick metric tibble.
}
\description{
Concentration-index gain metric
}
