% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_parsnip_formula.R
\name{ci_prediction_index}
\alias{ci_prediction_index}
\title{Concentration-index prediction metric}
\usage{
ci_prediction_index(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A yardstick metric tibble.
}
\description{
Concentration-index prediction metric
}
