% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_data_table_available.R
\name{.assert_data_table_available}
\alias{.assert_data_table_available}
\title{Assert that the data.table package is available}
\usage{
.assert_data_table_available()
}
\value{
Invisibly returns \code{NULL} if the check passes. Otherwise, throws an
error.
}
\description{
Assert that the data.table package is available
}
\details{
This function is used internally to ensure that the \code{data.table}
package is available before attempting to use it. If \code{data.table} is not
installed, an informative error message is thrown instructing the user to
install the package.
}
