% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_forest.R
\name{fitted.ci_forest}
\alias{fitted.ci_forest}
\title{Extract fitted values from a greedy concentration-index forest}
\usage{
\method{fitted}{ci_forest}(object, ...)
}
\arguments{
\item{object}{A fitted \code{ci_forest} object from \code{\link[=ci_forest]{ci_forest()}}.}

\item{...}{Currently ignored.}
}
\value{
A data frame with fitted values, response values, and weights.
}
\description{
Extract fitted values from a greedy concentration-index forest
}
