% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_kenya.R
\docType{data}
\name{kenya}
\alias{kenya}
\title{Simulated Kenya Child Survival Dataset}
\format{
A data frame with 20043 rows and 15 variables:
\describe{
\item{ wealth }{  numeric }
\item{ deadu5_num }{  integer }
\item{ quint }{  factor }
\item{ unskilled }{  factor }
\item{ male }{  factor }
\item{ birth }{  factor }
\item{ agemoth }{  factor }
\item{ rural }{  factor }
\item{ ed }{  factor }
\item{ ped }{  factor }
\item{ mocc }{  factor }
\item{ pocc }{  factor }
\item{ reg }{  factor }
\item{ sample_weight }{  numeric }
\item{ case_weight }{  integer }
}
}
\source{
Simulated in data-raw/simulated_kenya.R.
}
\description{
Simulated Kenya child survival dataset.
}
