% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_splitfun.R
\name{knit_print.ci_forest}
\alias{knit_print.ci_forest}
\title{Knit-print a greedy concentration-index forest}
\usage{
\method{knit_print}{ci_forest}(x, ...)
}
\arguments{
\item{x}{A fitted \code{ci_forest} object.}

\item{...}{Currently ignored.}
}
\value{
A \code{knitr} as-is output object.
}
\description{
Knit-print a greedy concentration-index forest
}
