% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_splitfun.R
\name{knit_print.ci_tree}
\alias{knit_print.ci_tree}
\title{Knit-print a greedy concentration-index tree}
\usage{
\method{knit_print}{ci_tree}(x, ...)
}
\arguments{
\item{x}{A fitted \code{ci_tree} object.}

\item{...}{Currently ignored.}
}
\value{
A \code{knitr} as-is output object.
}
\description{
Knit-print a greedy concentration-index tree
}
