% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_splitfun.R
\name{print.ci_forest}
\alias{print.ci_forest}
\title{Print a greedy concentration-index forest}
\usage{
\method{print}{ci_forest}(x, ..., digits = 3L)
}
\arguments{
\item{x}{A fitted \code{ci_forest} object.}

\item{...}{Currently ignored.}

\item{digits}{Number of digits used for printed numeric summaries.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a greedy concentration-index forest
}
