% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_splitfun.R
\name{print.ci_tree}
\alias{print.ci_tree}
\title{Print a greedy concentration-index tree}
\usage{
\method{print}{ci_tree}(x, ..., nodes = 10L, digits = 3L)
}
\arguments{
\item{x}{A fitted \code{ci_tree} object.}

\item{...}{Currently ignored.}

\item{nodes}{Number of terminal nodes to show in the console summary.}

\item{digits}{Number of digits used for printed numeric summaries.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a greedy concentration-index tree
}
