% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_parsnip_formula.R
\name{register_ineqtrees_decision_tree}
\alias{register_ineqtrees_decision_tree}
\title{Register the \code{decision_tree()} \code{ineqTrees} engine}
\usage{
register_ineqtrees_decision_tree()
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Register the \code{decision_tree()} \code{ineqTrees} engine
}
