% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_parsnip_formula.R
\name{register_ineqtrees_rand_forest}
\alias{register_ineqtrees_rand_forest}
\title{Register the \code{rand_forest()} \code{ineqTrees} engine}
\usage{
register_ineqtrees_rand_forest()
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Register the \code{rand_forest()} \code{ineqTrees} engine
}
