% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ci_tree_terminal_mean.R
\name{ci_select_best}
\alias{ci_select_best}
\title{Select the best tuning parameters}
\usage{
ci_select_best(x, metric = NULL, ...)
}
\arguments{
\item{x}{A tuning object returned by \code{\link[=tune_ci_tree]{tune_ci_tree()}} or \code{\link[=tune_ci_forest]{tune_ci_forest()}}.}

\item{metric}{Optional metric name used to filter the result.}

\item{...}{Reserved for future extensions.}
}
\value{
A one-row \code{data.table}.
}
\description{
Select the best tuning parameters
}
